+++ src/MCDRpost/mcdrpost/manager/command_manager.py
@@ -1,14 +1,16 @@
 from typing import TYPE_CHECKING
 
-from mcdreforged import CommandSource, GreedyText, InfoCommandSource, Integer, Literal, PluginServerInterface, RAction, \
+from mcdreforged import CommandContext, CommandSource, GreedyText, InfoCommandSource, Integer, Literal, \
+    PluginServerInterface, RAction, \
     RColor, RText, RTextList, RequirementNotMet, Text
 
 from mcdrpost.configuration import CommandPermission, Configuration
-from mcdrpost.constants import END_LINE
-from mcdrpost.utils.translation import Tags, tr
+from mcdrpost.constants import END_LINE, SIMPLE_HELP_MESSAGE
+from mcdrpost.utils import add_requirements
+from mcdrpost.utils.translation import TranslationKeys
 
 if TYPE_CHECKING:
-    from mcdrpost.manager.post_manager import PostManager  # noqa: F401
+    from mcdrpost.manager.post_manager import PostManager
 
 
 class CommandManager:
@@ -17,6 +19,8 @@ class CommandManager:
     def __init__(self, post_manager: "PostManager") -> None:
         self._post_manager: "PostManager" = post_manager
         self._server: PluginServerInterface = post_manager.server
+        self.logger = self._server.logger
+
         self._data_manager = post_manager.data_manager
 
         self._prefixes = ["!!po"]
@@ -38,10 +42,7 @@ class CommandManager:
             self._prefixes = self._config.command_prefixes
 
         for prefix in self._prefixes:
-            self._server.register_help_message(prefix, {
-                "en_us": "post/teleport weapon hands items",
-                "zh_cn": "传送/收寄副手物品",
-            })
+            self._server.register_help_message(prefix, SIMPLE_HELP_MESSAGE)
             self._server.register_command(
                 self.generate_command_node(prefix)
             )
@@ -56,46 +57,46 @@ class CommandManager:
             msgs_on_helper = RTextList(
                 RText(prefix + ' list orders', RColor.gray)
                 .c(RAction.suggest_command, f"{prefix} list orders")
-                .h(tr('hover')),
+                .h(TranslationKeys.hover.tr()),
 
-                RText(tr(Tags.help.hint_ls_orders) + END_LINE),
+                RText(TranslationKeys.help.hint_ls_orders.tr() + END_LINE),
             )
         if source.has_permission(3):
             # admin以上权限的添加信息
             msgs_on_admin = RTextList(
-                RText(prefix + tr(Tags.help.player_add), RColor.gray)
+                RText(prefix + TranslationKeys.help.player_add.tr(), RColor.gray)
                 .c(RAction.suggest_command, f"{prefix} player add ")
-                .h(tr('hover')), RText(f'{tr("help.hint_player_add")}\n'),
+                .h(TranslationKeys.hover.tr()), RText(f'{TranslationKeys.help.hint_player_add.tr()}\n'),
 
-                RText(prefix + tr(Tags.help.player_remove), RColor.gray)
+                RText(prefix + TranslationKeys.help.player_remove.tr(), RColor.gray)
                 .c(RAction.suggest_command, f"{prefix} player remove ")
-                .h(tr('hover')), RText(f'{tr("help.hint_player_remove")}\n'),
+                .h(TranslationKeys.hover.tr()), RText(f'{TranslationKeys.help.hint_player_remove.tr()}\n'),
             )
 
         source.reply(
             RTextList(
                 RText('--------- §3MCDRpost §r---------\n'),
-                RText(tr(Tags.desc) + END_LINE),
-                RText(tr(Tags.help.title) + END_LINE),
-                RText(prefix, RColor.gray).c(RAction.suggest_command, prefix).h(tr('hover')),
-                RText(f' | {tr(Tags.help.hint_help)}\n'),
-                RText(prefix + tr(Tags.help.p), RColor.gray).c(RAction.suggest_command, f"{prefix} post").h(
-                    tr(Tags.hover)),
-                RText(f'{tr(Tags.help.hint_p)}\n'),
-                RText(prefix + ' rl', RColor.gray).c(RAction.suggest_command, f"{prefix} receive_list").h(tr('hover')),
-                RText(f'{tr(Tags.help.hint_rl)}\n'),
-                RText(prefix + tr('help.r'), RColor.gray).c(RAction.suggest_command, f"{prefix} receive").h(
-                    tr('hover')),
-                RText(f'{tr(Tags.help.hint_r)}\n'),
+                RText(TranslationKeys.desc.tr() + END_LINE),
+                RText(TranslationKeys.help.title.tr() + END_LINE),
+                RText(prefix, RColor.gray).c(RAction.suggest_command, prefix).h(TranslationKeys.hover.tr()),
+                RText(f' | {TranslationKeys.help.hint_help.tr()}\n'),
+                RText(prefix + TranslationKeys.help.p.tr(), RColor.gray).c(RAction.suggest_command, f"{prefix} post").h(
+                    TranslationKeys.hover.tr()),
+                RText(f'{TranslationKeys.help.hint_p.tr()}\n'),
+                RText(prefix + ' rl', RColor.gray).c(RAction.suggest_command, f"{prefix} receive_list").h(TranslationKeys.hover.tr()),
+                RText(f'{TranslationKeys.help.hint_rl.tr()}\n'),
+                RText(prefix + TranslationKeys.help.r.tr(), RColor.gray).c(RAction.suggest_command, f"{prefix} receive").h(
+                    TranslationKeys.hover.tr()),
+                RText(f'{TranslationKeys.help.hint_r.tr()}\n'),
                 RText(prefix + ' pl', RColor.gray)
-                .c(RAction.suggest_command, f"{prefix} post_list").h(tr('hover')),
-                RText(f'{tr(Tags.help.hint_pl)}\n'),
-                RText(prefix + tr(Tags.help.c), RColor.gray)
-                .c(RAction.suggest_command, f"{prefix} cancel").h(tr('hover')),
-                RText(f'{tr(Tags.help.hint_c)}\n'),
+                .c(RAction.suggest_command, f"{prefix} post_list").h(TranslationKeys.hover.tr()),
+                RText(f'{TranslationKeys.help.hint_pl.tr()}\n'),
+                RText(prefix + TranslationKeys.help.c.tr(), RColor.gray)
+                .c(RAction.suggest_command, f"{prefix} cancel").h(TranslationKeys.hover.tr()),
+                RText(f'{TranslationKeys.help.hint_c.tr()}\n'),
                 RText(prefix + ' ls players', RColor.gray)
-                .c(RAction.suggest_command, f"{prefix} list players").h(tr('hover')),
-                RText(f'{tr(Tags.help.hint_ls_players)}\n'),
+                .c(RAction.suggest_command, f"{prefix} list players").h(TranslationKeys.hover.tr()),
+                RText(f'{TranslationKeys.help.hint_ls_players.tr()}\n'),
                 msgs_on_helper,
                 msgs_on_admin,
                 RText("§a『别名 Alias』§r\n"),
@@ -115,7 +116,7 @@ class CommandManager:
         )
 
         if not post_list:
-            src.reply(tr(Tags.no_post_orders))
+            src.reply(TranslationKeys.no_post_orders.tr())
             return
 
         msg = ""
@@ -130,7 +131,7 @@ class CommandManager:
             '-------------------------------------------\n'
             '{2}'
             '===========================================\n'
-            .format(tr(Tags.list_post_orders_title), msg, tr(Tags.hint_cancel))
+            .format(TranslationKeys.list_post_orders_title.tr(), msg, TranslationKeys.hint_cancel.tr())
         )
 
     def output_receive_list(self, src: InfoCommandSource) -> None:
@@ -140,7 +141,7 @@ class CommandManager:
         )
 
         if not receive_list:
-            src.reply(tr(Tags.no_receive_orders))
+            src.reply(TranslationKeys.no_receive_orders.tr())
             return
 
         msg = ""
@@ -155,7 +156,7 @@ class CommandManager:
             '-------------------------------------------\n'
             '{2}'
             '===========================================\n'
-            .format(tr(Tags.list_receive_orders_title), msg, tr(Tags.hint_order_receive))
+            .format(TranslationKeys.list_receive_orders_title.tr(), msg, TranslationKeys.hint_order_receive.tr())
         )
 
     def output_all_orders(self, src: InfoCommandSource) -> None:
@@ -163,7 +164,7 @@ class CommandManager:
         all_orders = self._data_manager.get_orders()
 
         if not all_orders:
-            src.reply(tr(Tags.no_orders))
+            src.reply(TranslationKeys.no_orders.tr())
             return
 
         msg = ""
@@ -176,27 +177,40 @@ class CommandManager:
             '{0}\n'
             '{1}\n'
             '===========================================\n'
-            .format(tr(Tags.list_orders_title), msg)
+            .format(TranslationKeys.list_orders_title.tr(), msg)
         )
 
     def receive(self, src: InfoCommandSource, order_id):
         if self._post_manager.receive(src, order_id, 'receive'):
-            src.reply(tr(Tags.receive_success, order_id))
+            src.reply(TranslationKeys.receive_success.tr(order_id))
 
     def cancel(self, src: InfoCommandSource, order_id):
         if self._post_manager.receive(src, order_id, 'cancel'):
-            src.reply(tr(Tags.cancel_success, order_id))
+            src.reply(TranslationKeys.cancel_success.tr(order_id))
+
+    def add_player(self, src: CommandSource, ctx: CommandContext):
+        player = ctx['player_id']
+        if not self._data_manager.add_player(player):
+            src.reply(TranslationKeys.has_player.tr(player))
+            return
+
+        src.reply(TranslationKeys.login_success.tr(player))
+        self.logger.info(TranslationKeys.login_log.tr(player))
+
+    def remove_player(self, src: CommandSource, ctx: CommandContext):
+        player = ctx['player_id']
+        if not self._data_manager.remove_player(player):
+            src.reply(TranslationKeys.cannot_del_player.tr(player))
+            return
+
+        src.reply(TranslationKeys.del_player_success.tr(player))
+        self.logger.info(TranslationKeys.del_player_log.tr(player))
 
     # nodes
-
     def gen_post_node(self, node_name: str) -> Literal:
-        return (
+        return add_requirements(
             Literal(node_name).
-            requires(lambda src: src.is_player and src.has_permission(self._perm.post)).
-            on_error(RequirementNotMet, lambda src: src.reply(
-                tr(Tags.no_permission if src.is_player else Tags.only_for_player)
-            ), handled=True).
-            runs(lambda src: src.reply(tr(Tags.no_input_receiver))).
+            runs(lambda src: src.reply(TranslationKeys.no_input_receiver.tr())).
             then(
                 Text('receiver').
                 suggests(self._data_manager.get_players).
@@ -205,27 +219,22 @@ class CommandManager:
                     GreedyText('comment').
                     runs(lambda src, ctx: self._post_manager.post(src, ctx['receiver'], ctx['comment']))
                 )
-            )
+            ),
+            permission=self._perm.post,
+            require_player=True
         )
 
     def gen_post_list_node(self, node_name: str) -> Literal:
-        return (
-            Literal(node_name).
-            requires(lambda src: src.is_player and src.has_permission(self._perm.post)).
-            on_error(RequirementNotMet, lambda src: src.reply(
-                tr(Tags.no_permission if not src.is_player else Tags.only_for_player)
-            ), handled=True).
-            runs(lambda src: self.output_post_list(src))
+        return add_requirements(
+            Literal(node_name).runs(lambda src: self.output_post_list(src)),
+            permission=self._perm.post,
+            require_player=True
         )
 
     def gen_receive_node(self, node_name: str) -> Literal:
-        return (
+        return add_requirements(
             Literal(node_name).
-            requires(lambda src: src.is_player and src.has_permission(self._perm.receive)).
-            on_error(RequirementNotMet, lambda src: src.reply(
-                tr(Tags.no_permission if src.is_player else Tags.only_for_player)
-            ), handled=True).
-            runs(lambda src: src.reply(tr(Tags.no_input_receive_orderid))).
+            runs(lambda src: src.reply(TranslationKeys.no_input_receive_orderid.tr())).
             then(
                 Integer('orderid').
                 suggests(
@@ -235,27 +244,22 @@ class CommandManager:
                     ]
                 ).
                 runs(lambda src, ctx: self.receive(src, ctx['orderid']))
-            )
+            ),
+            permission=self._perm.receive,
+            require_player=True
         )
 
     def gen_receive_list_node(self, node_name: str) -> Literal:
-        return (
-            Literal(node_name).
-            requires(lambda src: src.is_player and src.has_permission(self._perm.receive)).
-            on_error(RequirementNotMet, lambda src: src.reply(
-                tr(Tags.no_permission if src.is_player else Tags.only_for_player)
-            ), handled=True).
-            runs(lambda src: self.output_receive_list(src))
+        return add_requirements(
+            Literal(node_name).runs(lambda src: self.output_receive_list(src)),
+            permission=self._perm.receive,
+            require_player=True
         )
 
     def gen_cancel_node(self, node_name: str) -> Literal:
-        return (
+        return add_requirements(
             Literal(node_name).
-            requires(lambda src: src.is_player and src.has_permission(self._perm.cancel)).
-            on_error(RequirementNotMet, lambda src: src.reply(
-                tr(Tags.no_permission if src.is_player else Tags.only_for_player)
-            ), handled=True).
-            runs(lambda src: src.reply(tr(Tags.no_input_cancel_orderid))).
+            runs(lambda src: src.reply(TranslationKeys.no_input_cancel_orderid.tr())).
             then(
                 Integer('orderid').
                 suggests(
@@ -265,40 +269,40 @@ class CommandManager:
                     ]
                 ).
                 runs(lambda src, ctx: self.cancel(src, ctx['orderid']))
-            )
+            ),
+            permission=self._perm.cancel,
+            require_player=True
         )
 
     def gen_list_node(self, node_name: str) -> Literal:
         return (
             Literal(node_name).
-            runs(lambda src: src.reply(tr(Tags.command_incomplete))).
+            runs(lambda src: src.reply(TranslationKeys.command_incomplete.tr())).
             then(
                 Literal('players').
                 requires(lambda src: src.has_permission(self._perm.list_player)).
                 runs(lambda src: src.reply(
-                    tr(Tags.list_player_title) + str(self._data_manager.get_players())
+                    TranslationKeys.list_player_title.tr() + str(self._data_manager.get_players())
                 ))
             ).
             then(
                 Literal('orders').
                 requires(lambda src: src.has_permission(self._perm.list_orders)).
-                on_error(RequirementNotMet, lambda src: src.reply(tr(Tags.no_permission)), handled=True).
+                on_error(RequirementNotMet, lambda src: src.reply(TranslationKeys.no_permission.tr()), handled=True).
                 runs(lambda src: self.output_all_orders(src))
             ).
             then(
-                Literal('receive').
-                requires(lambda src: src.is_player and src.has_permission(self._perm.receive)).
-                on_error(RequirementNotMet, lambda src: src.reply(
-                    tr(Tags.no_permission if src.is_player else Tags.only_for_player)
-                ), handled=True).
-                runs(lambda src: self.output_receive_list(src))
+                add_requirements(
+                    Literal('receive').runs(lambda src: self.output_receive_list(src)),
+                    permission=self._perm.receive,
+                    require_player=True
+                )
             ).then(
-                Literal('post').
-                requires(lambda src: src.is_player and src.has_permission(self._perm.post)).
-                on_error(RequirementNotMet, lambda src: src.reply(
-                    tr(Tags.no_permission if src.is_player else Tags.only_for_player)
-                ), handled=True).
-                runs(lambda src: self.output_post_list(src))
+                add_requirements(
+                    Literal('post').runs(lambda src: self.output_post_list(src)),
+                    permission=self._perm.post,
+                    require_player=True
+                )
             )
         )
 
@@ -306,33 +310,34 @@ class CommandManager:
         return (
             Literal(node_name).
             requires(lambda src: src.has_permission(self._perm.player)).
-            on_error(RequirementNotMet, lambda src: src.reply(tr(Tags.no_permission)), handled=True).
-            runs(lambda src: src.reply(tr(Tags.command_incomplete))).
+            on_error(RequirementNotMet, lambda src: src.reply(TranslationKeys.no_permission.tr()), handled=True).
+            runs(lambda src: src.reply(TranslationKeys.command_incomplete.tr())).
             then(
                 Literal('add').
-                runs(lambda src: src.reply(tr(Tags.command_incomplete))).
-                then(
-                    Text('player_id').
-                    runs(lambda src, ctx: self._data_manager.add_player(ctx['player_id']))
-                )
+                runs(lambda src: src.reply(TranslationKeys.command_incomplete.tr())).
+                then(Text('player_id').runs(self.add_player))
             ).
             then(
                 Literal('remove').
-                runs(lambda src: src.reply(tr(Tags.command_incomplete))).
+                runs(lambda src: src.reply(TranslationKeys.command_incomplete.tr())).
                 then(
                     Text('player_id').
                     suggests(self._data_manager.get_players).
-                    runs(lambda src, ctx: self._data_manager.remove_player(ctx['player_id']))
+                    runs(self.remove_player)
                 )
             )
         )
 
     def gen_reload_node(self, prefix) -> Literal:
+        def reload(src: CommandSource):
+            self._post_manager.reload()
+            src.reply(TranslationKeys.reload_success.tr())
+
         return (
             Literal(prefix)
             .requires(lambda src: src.has_permission(self._perm.reload))
-            .on_error(RequirementNotMet, lambda src: src.reply(tr(Tags.no_permission)), handled=True)
-            .runs(lambda src: (self._post_manager.reload(), src.reply(tr(Tags.reload_success))))
+            .on_error(RequirementNotMet, lambda src: src.reply(TranslationKeys.no_permission.tr()), handled=True)
+            .runs(reload)
         )
 
     def generate_command_node(self, prefix: str) -> Literal:
@@ -340,7 +345,7 @@ class CommandManager:
         return (
             Literal(prefix).
             requires(lambda src: src.has_permission(self._perm.root)).
-            on_error(RequirementNotMet, lambda src: src.reply(tr(Tags.no_permission)), handled=True).
+            on_error(RequirementNotMet, lambda src: src.reply(TranslationKeys.no_permission.tr()), handled=True).
             runs(lambda src: self.output_help_message(src, prefix)).
             # 下面的一行就是一条命令，多个 then 意味着别名/缩写
             then(self.gen_post_node('p')).then(self.gen_post_node('post')).