+++ src/MCDRpost/mcdrpost/manager/config_manager.py
@@ -4,10 +4,10 @@ from mcdreforged import PluginServerInterface
 
 from mcdrpost import constants
 from mcdrpost.configuration import Configuration
-from mcdrpost.utils.translation import Tags, tr
+from mcdrpost.utils.translation import TranslationKeys
 
 if TYPE_CHECKING:
-    from mcdrpost.manager.post_manager import PostManager  # noqa: F401
+    from mcdrpost.manager.post_manager import PostManager
 
 
 class ConfigurationManager:
@@ -16,11 +16,15 @@ class ConfigurationManager:
     def __init__(self, post_manager: "PostManager") -> None:
         self._post_manager: "PostManager" = post_manager
         self._server: PluginServerInterface = post_manager.server
-        self._configuration: Configuration | None = None
+        self._configuration: Configuration = self._server.load_config_simple(
+            constants.CONFIG_FILE_NAME,
+            target_class=Configuration,
+            file_format=constants.CONFIG_FILE_TYPE,
+        )
 
     def reload(self) -> None:
         """(重新)加载配置文件"""
-        self._server.logger.info(tr(Tags.config.load))
+        self._server.logger.info(TranslationKeys.config.load.tr())
         self._configuration = self._server.load_config_simple(
             constants.CONFIG_FILE_NAME,
             target_class=Configuration,