+++ src/MCDRpost/mcdrpost/manager/version_manager.py
@@ -6,6 +6,7 @@ from mcdreforged import PluginServerInterface
 from mcdrpost.data_structure import Item
 from mcdrpost.environment import Environment
 from mcdrpost.version_handler.abstract_version_handler import AbstractVersionHandler
+from mcdrpost.version_handler.sound_player.abstract_sound_player import AbstractSoundPlayer
 
 Checker = Callable[[Environment], bool]
 
@@ -78,6 +79,8 @@ class VersionManager:
             player (str): 玩家名
             item (str): 物品字符串
         """
+        if self._handler is None:
+            raise RuntimeError("version handler is not initialized")
         self._handler.replace(player, item)
 
     def get_offhand_item(self, player: str) -> Item:
@@ -86,4 +89,12 @@ class VersionManager:
         Args:
             player (str): 玩家名
         """
+        if self._handler is None:
+            raise RuntimeError("version handler is not initialized")
         return self._handler.get_offhand_item(player)
+
+    @property
+    def play_sound(self) -> AbstractSoundPlayer:
+        if self._handler is None:
+            raise RuntimeError("version handler is not initialized")
+        return self._handler.play_sound